Address PR review: rename stack module to stackx_*; enable strict warnings
Review feedback on PR #1:

1. Rename the stack data structure for consistency and to avoid the POSIX
   reserved type collision.
     - stack_t collides with POSIX stack_t (declared in <signal.h> for
       sigaltstack), which newlib defines on ESP-IDF, causing a clash.
     - Renamed type clc_stack_t -> stackx_t and all API functions
       stack_* -> stackx_*, matching the existing socketx/stringx 'x' suffix
       convention so the type prefix and function prefix agree.
     - Renamed inc/stack.h -> inc/stackx.h and src/stack.c -> src/stackx.c.
     - Updated all callers (bst_redblack.c, tests, common_lib_c.h, CMake).

2. Enable the highest practical compiler diagnostics on the library target.
     - Add COMMON_LIB_C_WARNINGS_AS_ERRORS option (default ON).
     - GCC/Clang: -Wall -Wextra -Wpedantic plus -Wshadow, -Wcast-qual,
       -Wcast-align, -Wundef, -Wwrite-strings, -Wpointer-arith,
       -Wstrict-prototypes, -Wmissing-prototypes, -Wmissing-declarations,
       -Wredundant-decls, -Wdouble-promotion, -Wformat=2, -Wnull-dereference,
       -Werror. MSVC: /W4 /permissive- /WX. Applied PRIVATE.
     - Fixed all resulting warnings: () -> (void) prototypes (base64.h,
       circular_list.{h,c}, socketx.h), cast-qual in span_from_str_literal,
       and removed the unused node_create() in bst_redblack.c.

+++ inc/span.h
@@ -49,7 +49,7 @@ int span_set(span_t span, uint32_t position, uint8_t value);
 
 #define span_from_str_literal(X) \
   (span_t){ \
-      .ptr = (uint8_t*)X, \
+      .ptr = (uint8_t*)(uintptr_t)(X), \
       .length = strlitlen(X) \
   }
